moved some time/sync funcs/consts out from syscall_utils
